ci: disable PR triggers in Azure Pipelines config
Whenever a PR is opened and commits are force-/pushed to the corresponding
branch, up to present the behavior has been that two sets of Azure Pipelines
builds are run: one is caused by the default PR trigger and the other is caused
by the default commits trigger. With the changes in this commit, the goal is to
retain the behavior of builds running for every commit to any branch (including
PR branches) but to disable the PR trigger. That will be helpful when our Azure
Pipelines cofig is soon updated with a build matrix that includes Node v12.x
and v13.x, i.e. instead of 18 builds any time a PR is opened or updated there
will be 9.

This change may not achieve the desired outcome; if it does not then it will be
undone.